Remove legacy Ray code. (#3121)
* Remove legacy Ray code.

* Fix cmake and simplify monitor.

* Fix linting

* Updates

* Fix

* Implement some methods.

* Remove more plasma manager references.

* Fix

* Linting

* Fix

* Fix

* Make sure class IDs are strings.

* Some path fixes

* Fix

* Path fixes and update arrow

* Fixes.

* linting

* Fixes

* Java fixes

* Some java fixes

* TaskLanguage -> Language

* Minor

* Fix python test and remove unused method signature.

* Fix java tests

* Fix jenkins tests

* Remove commented out code.

+++ python/ray/test/cluster_utils.py
@@ -44,7 +44,6 @@ class Cluster(object):
 
         All nodes are by default started with the following settings:
             cleanup=True,
-            use_raylet=True,
             resources={"CPU": 1},
             object_store_memory=100 * (2**20) # 100 MB
 
@@ -55,12 +54,13 @@ class Cluster(object):
         Returns:
             Node object of the added Ray node.
         """
-        node_kwargs = dict(
-            cleanup=True,
-            use_raylet=True,
-            resources={"CPU": 1},
-            object_store_memory=100 * (2**20)  # 100 MB
-        )
+        node_kwargs = {
+            "cleanup": True,
+            "resources": {
+                "CPU": 1
+            },
+            "object_store_memory": 100 * (2**20)  # 100 MB
+        }
         node_kwargs.update(override_kwargs)
 
         if self.head_node is None:
@@ -179,7 +179,9 @@ class Node(object):
         for process_name, process_list in self.process_dict.items():
             logger.info("Killing all {}(s)".format(process_name))
             for process in process_list:
-                process.kill()
+                # Kill the process if it is still alive.
+                if process.poll() is None:
+                    process.kill()
 
         for process_name, process_list in self.process_dict.items():
             logger.info("Waiting all {}(s)".format(process_name))
